Fleshed out api functions in content.py
Added history dict to Content, which maps an ingest_uuid to a state
(QUEUED, INGESTING, INGESTED, CANCELLED = range(4)). Added code to return
this information to the client, and also allow the client to clear the
history. Client can now also cancel a dcp ingest (as long as it has not
already started ingesting)